requirements.txt
torch==1.7.1 + cu110
numpy==1.19.4
opencv_python==3.4.4.19
scipy==1.1.0
pillow ==8.0.1
scikit_learn==0.24.2
matplotlib==3.4.3 
imageio==2.16.0
pytorch-msssim==0.2.1

